// src/utils/markdown/simple-markdown.d.ts
export interface Capture extends Array<string> {
    index?: number,
    input?: string,
}
type Attr = string | number | boolean | null | undefined;
export type TableAlignment = "right" | "center" | "left" | null;

export interface SingleASTNode {
    [x: string]: BlockStyleTypes;
    astNode: BlockStyleTypes;
    type: string,
    [prop: string]: any,
}

export interface UnTypedASTNode {
    [prop: string]: any
}

export type ASTNode = SingleASTNode | Array<SingleASTNode>;

export interface State {
    key?: string | number | undefined;
    inline?: boolean | undefined;
    [prop: string]: any,
}
export type OptionalState = State | null | undefined;

export type ReactElement = React.ReactElement<any>;
export type ReactElements = React.ReactNode;

export interface MatchFunction {
    (source: string, state: State, prevCapture: string): Capture | null,
    regex?: RegExp,
}

export type Parser = (
    source: string,
    state?: OptionalState,
) => Array<SingleASTNode>;

export type ParseFunction = (
    capture: Capture,
    nestedParse: Parser,
    state: State,
) => (UnTypedASTNode | ASTNode);

export type SingleNodeParseFunction = (
    capture: Capture,
    nestedParse: Parser,
    state: State,
) => UnTypedASTNode;

export type Output<Result> = (
    node: ASTNode,
    state?: OptionalState
) => Result;

export type RefiningNodeOutput<Input, Result extends Input> = (
    node: SingleASTNode,
    nestedOutput: Output<Input>,
    state: State
) => Result;

export type NodeOutput<Result> = RefiningNodeOutput<Result, Result>;

export type ArrayNodeOutput<Result> = (
    node: Array<SingleASTNode>,
    nestedOutput: Output<Result>,
    state: State
) => Result;

export type ReactOutput = Output<ReactElements>;
export type ReactNodeOutput = NodeOutput<ReactElements>;
export type HtmlOutput = Output<string>;
export type HtmlNodeOutput = NodeOutput<string>;

export interface ParserRule {
    readonly order: number,
    readonly match: MatchFunction,
    readonly quality?: (capture: Capture, state: State, prevCapture: string) => number,
    readonly parse: ParseFunction,
}
// ...
export interface ArrayRule {
    readonly react?: ArrayNodeOutput<ReactElements>,
    readonly html?: ArrayNodeOutput<string>,
    readonly [other: string]: ArrayNodeOutput<any> | undefined,
}
// ...
export interface ParserRules {
    readonly Array?: ArrayRule,
    readonly [type: string]: ParserRule | /* only for Array: */ ArrayRule | undefined,
}
// ...
export const parserFor: (rules: ParserRules, defaultState?: OptionalState) => Parser;

export const outputFor: OutputFor;

export const inlineRegex: (regex: RegExp) => MatchFunction;
export const blockRegex: (regex: RegExp) => MatchFunction;
export const anyScopeRegex: (regex: RegExp) => MatchFunction;
export const parseInline: (parse: Parser, content: string, state: State) => ASTNode;
export const parseBlock: (parse: Parser, content: string, state: State) => ASTNode;
// ...
export const defaultHtmlOutput: HtmlOutput;